+++ src/tewi/app.py
@@ -31,6 +31,8 @@ from transmission_rpc import Client
 from transmission_rpc.error import TransmissionError
 from transmission_rpc.session import Session, SessionStats
 
+from rich.text import Text
+
 from textual import on, work
 from textual.app import App, ComposeResult
 from textual.binding import Binding
@@ -38,8 +40,7 @@ from textual.containers import Grid, ScrollableContainer, Horizontal, Container
 from textual.message import Message
 from textual.reactive import reactive
 from textual.screen import ModalScreen
-from textual.widgets import Static, Label, ProgressBar, DataTable, ContentSwitcher, TabbedContent, TabPane, TextArea, \
-        ListView, ListItem
+from textual.widgets import Static, Label, ProgressBar, DataTable, ContentSwitcher, TabbedContent, TabPane, TextArea
 
 from geoip2fast import GeoIP2Fast
 
@@ -81,6 +82,8 @@ class PageState(NamedTuple):
 class SortOrder(NamedTuple):
     id: str
     name: str
+    key_asc: str
+    key_desc: str
     sort_func: None
 
 
@@ -93,14 +96,20 @@ class TransmissionSession(NamedTuple):
     torrents_complete_size: int
     torrents_total_size: int
     sort_order: SortOrder
+    sort_order_asc: bool
 
 
 sort_orders = [
-        SortOrder('name', 'Name', lambda t: t.name.lower()),
-        SortOrder('status', 'Status', lambda t: t.status),
-        SortOrder('size', 'Size', lambda t: t.total_size),
-        SortOrder('progress', 'Progress', lambda t: t.percent_done),
-        SortOrder('ratio', 'Ratio', lambda t: t.ratio),
+        SortOrder('name', 'Name', 'n', 'N',
+                  lambda t: t.name.lower()),
+        SortOrder('status', 'Status', 's', 'S',
+                  lambda t: t.status),
+        SortOrder('size', 'Size', 'z', 'Z',
+                  lambda t: t.total_size),
+        SortOrder('progress', 'Progress', 'p', 'P',
+                  lambda t: t.percent_done),
+        SortOrder('ratio', 'Ratio', 'r', 'R',
+                  lambda t: t.ratio),
         ]
 
 
@@ -452,35 +461,35 @@ class SortOrderWidget(Static):
             ]
 
     def compose(self) -> ComposeResult:
-        yield SortOrderListView()
+        yield DataTable(cursor_type="none",
+                        zebra_stripes=True)
 
     def on_mount(self) -> None:
         self.border_title = 'Sort order'
-        self.border_subtitle = '[Enter,L] Select / [X] Close'
+        self.border_subtitle = '[X] Close'
 
-        list_view = self.query_one(SortOrderListView)
+        table = self.query_one(DataTable)
+        table.add_columns("Order", "Key (ASC | DESC)")
 
-        for order in sort_orders:
-            list_view.append(ListItem(Label(order.name), id=order.id))
+        for o in sort_orders:
+            table.add_row(o.name,
+                          Text(str(f'   {o.key_asc} | {o.key_desc}'), justify="center"))
 
-    @on(ListView.Selected)
-    def handle_selection(self, event: ListView.Selected) -> None:
-        order = next(x for x in sort_orders if x.id == event.item.id)
+            b = Binding(o.key_asc, f"select_order('{o.id}', True)")
+            self._bindings._add_binding(b)
 
-        self.post_message(MainApp.SortOrderSelected(order))
+            b = Binding(o.key_desc, f"select_order('{o.id}', False)")
+            self._bindings._add_binding(b)
+
+    def action_select_order(self, sort_id, is_asc):
+        order = next(x for x in sort_orders if x.id == sort_id)
+
+        self.post_message(MainApp.SortOrderSelected(order, is_asc))
 
         self.parent.dismiss(False)
 
     def action_close(self) -> None:
         self.parent.dismiss(False)
-
-
-class SortOrderListView(ListView):
-    BINDINGS = [
-            Binding("k", "cursor_up", "Cursor up"),
-            Binding("j", "cursor_down", "Cursor down"),
-            Binding("l", "select_cursor", "Select"),
-            ]
 
 
 # Core UI panels
@@ -578,7 +587,9 @@ class StatePanel(Static):
                 self.r_stats_size = f'Size: {complete_size}'
 
             sort_order = new_r_tsession.sort_order.name
-            self.r_sort = f'Sort: {sort_order}'
+            sort_order_asc = new_r_tsession.sort_order_asc
+            sort_arrow = '' if sort_order_asc else '↑'
+            self.r_sort = f'Sort: {sort_order}{sort_arrow}'
 
             self.r_upload_speed = session_stats.upload_speed
             self.r_download_speed = session_stats.download_speed
@@ -1462,9 +1473,10 @@ class MainApp(App):
             self.value = value
 
     class SortOrderSelected(Message):
-        def __init__(self, order: str) -> None:
+        def __init__(self, order: str, is_asc: bool) -> None:
             super().__init__()
             self.order = order
+            self.is_asc = is_asc
 
     class PageChanged(Message):
         def __init__(self, state: PageState) -> None:
@@ -1517,6 +1529,7 @@ class MainApp(App):
         self.transmission_version = self.client.get_session().version
 
         self.sort_order = sort_orders[0]
+        self.sort_order_asc = True
 
     @log_time
     def compose(self) -> ComposeResult:
@@ -1567,9 +1580,11 @@ class MainApp(App):
                 torrents_complete_size=torrents_complete_size,
                 torrents_total_size=torrents_total_size,
                 sort_order=self.sort_order,
+                sort_order_asc=self.sort_order_asc,
                 )
 
-        torrents.sort(key=self.sort_order.sort_func)
+        torrents.sort(key=self.sort_order.sort_func,
+                      reverse=not self.sort_order_asc)
 
         self.log(vars(session))
         self.log(vars(session_stats))
@@ -1654,8 +1669,11 @@ class MainApp(App):
     @on(SortOrderSelected)
     def handle_sort_order_selected(self, event: SortOrderSelected) -> None:
         self.sort_order = event.order
+        self.sort_order_asc = event.is_asc
+
+        direction = 'ASC' if event.is_asc else 'DESC'
         self.post_message(MainApp.Notification(
-            f"Selected sort order: {event.order.name}"))
+            f"Selected sort order: {event.order.name} {direction}"))
 
     @log_time
     @on(PageChanged)